+++ src/inspect_sandbox_tools/src/inspect_sandbox_tools/_remote_tools/_bash_session/json_rpc_methods.py
@@ -20,7 +20,7 @@ _can_switch_user = os.getuid() == 0
 async def bash_session_new_session(params: NewSessionParams) -> NewSessionResult:
     return NewSessionResult(
         session_name=await controller.new_session(
-            user=params.user, can_switch_user=_can_switch_user
+            user=params.user or params.run_as, can_switch_user=_can_switch_user
         )
     )